+++ CHANGELOG.md
@@ -15,6 +15,7 @@ active session. This should not happen anymore. (Did not happen on Xenial but on
 ### Added
 - More informative logging output when a file included in the configuration could not be found.
 - Single line definition of requirements instead of lists get detected and an appropriate hint is shown in the log.
+- Local host stats monitoring and display of it in interactive cli.
 
 ## [2.2.0] - 19.06.2019